requirements.txt
streamlit
openai-whisper
langchain
langchain-community
langchain-core
langchain-groq
python-dotenv
pydantic
gTTS